Hotfix de campo horas en los proyectos

+++ isc.time.report.be/isc.time.report.be.domain/Models/Response/Projects/GetAllProjectsResponse.cs
@@ -22,7 +22,7 @@ namespace isc.time.report.be.domain.Models.Response.Projects
         public DateTime? ActualStartDate { get; set; }
         public DateTime? ActualEndDate { get; set; }
         public decimal? Budget { get; set; }
-        public int Hours { get; set; }
+        public decimal Hours { get; set; }
         public bool? Status { get; set; }
     }
 }
